Update src/GSMClient.cpp
Move set of GSMClient state to CLIENT_STATE_IDLE to the top of the stop() method so that the state is reset even if the _socket value is -1.  This is necessary because the state can be left in CLIENT_STATE_WAIT_CREATE_SOCKET_RESPONSE state with _socket = -1 if the AT+USOCR= command hangs.